# Make dataloaders slightly faster by using a custom transpose function.