The new function used for transposing the batches when 'collating' in the dataloader is ~6x. This makes dataloader in general slightly faster ~3/4% in a typical example.

``` r
library(purrr)
library(collapse)
#> Warning: package 'collapse' was built under R version 4.1.2
#> collapse 1.7.5, see ?`collapse-package` or ?`collapse-documentation`
#> 
#> Attaching package: 'collapse'
#> The following object is masked from 'package:stats':
#> 
#>     D

batch <- purrr::rerun(1000, x = array(1, c(32,32,3)), y = runif(5))
transpose2 <- torch:::transpose2

bench::mark(
  old = lapply(setNames(1:2, names(batch[[1]])), function(i) {
    lapply(batch, function(x) x[[i]])
  }),
  new = transpose2(batch),
  purrr = transpose(batch),
  base = do.call(Map, c(f = list, batch)),
  collapase = t_list(batch)
)
#> # A tibble: 5 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 old           928µs   1.02ms      966.      94KB     51.8
#> 2 new           123µs 139.42µs     6958.    18.2KB     44.3
#> 3 purrr         205µs 226.97µs     4301.    18.1KB     52.4
#> 4 base          822µs 927.39µs     1034.    70.8KB     50.0
#> 5 collapase     291µs 331.25µs     2941.    40.9KB     21.3
```

<sup>Created on 2022-02-11 by the [reprex package](https://reprex.tidyverse.org) (v2.0.1)</sup>